# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "KYS PORTAL",
    "description": "KYS Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "103.224.246.103:3004",
  "basePath": "/",
...
  "paths": {
...
    "/news": {
      "post": {
        "tags": [
          "news"
        ],
        "description": "Create new news in system",
        "parameters": [
          {
            "name": "attachment",
            "in": "formData",
            "description": "attachment",
            "type": "file"
          },
          {
            "name": "newsName",
            "in": "formData",
            "description": "newsName",
            "type": "string"
          },
          {
            "name": "companyName",
            "in": "formData",
            "description": "companyName",
            "type": "string"
          },
          {
            "name": "newsDetails",
            "in": "formData",
            "description": "newsDetails",
            "type": "string"
          },
          {
            "name": "active",
            "in": "formData",
            "description": "active",
            "type": "boolean"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New news is created",
            "schema": {
              "$ref": "#/definitions/news"
            }
          }
        }
      }
    },
    "/news/deleteById": {
      "get": {
        "tags": [
          "news"
        ],
        "description": "Remove news in system",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New news is created",
            "schema": {
              "$ref": "#/definitions/news"
            }
          }
        }
      }
    },
    "/news/update": {
      "post": {
        "tags": [
          "news"
        ],
        "description": "Create new news in system",
        "parameters": [
          {
            "name": "attachment",
            "in": "formData",
            "description": "attachment",
            "type": "file"
          },
          {
            "name": "_id",
            "in": "formData",
            "description": "id",
            "type": "string"
          },
          {
            "name": "newsName",
            "in": "formData",
            "description": "newsName",
            "type": "string"
          },
          {
            "name": "companyName",
            "in": "formData",
            "description": "companyName",
            "type": "string"
          },
          {
            "name": "newsDetails",
            "in": "formData",
            "description": "newsDetails",
            "type": "string"
          },
          {
            "name": "active",
            "in": "formData",
            "description": "active",
            "type": "boolean"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New news is created",
            "schema": {
              "$ref": "#/definitions/news"
            }
          }
        }
      }
    },
...
    "/plan": {
      "post": {
        "tags": [
          "Plans"
        ],
        "description": "Create new plan in system",
        "parameters": [
          {
            "name": "plan",
            "in": "body",
            "description": "plan that we want to create",
            "schema": {
              "$ref": "#/definitions/Plan"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New plan is created",
            "schema": {
              "$ref": "#/definitions/Plan"
            }
          }
        }
      }
    } ,
...
    "/plan/update": {
      "post": {
        "tags": [
          "Plans"
        ],
        "description": "update plan in system",
        "parameters": [
          {
            "name": "plan",
            "in": "body",
            "description": "plan that we want to create",
            "schema": {
              "$ref": "#/definitions/UpdatePlan"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New plan is created",
            "schema": {
              "$ref": "#/definitions/UpdatePlan"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "news": {
      "required": [
        "newsName"
      ],
      "properties": {
        "newsName": {
          "type": "string"
        },
        "companyName": {
          "type": "string"
        },
        "newsDetails": {
          "type": "string"
        },
        "active": {
          "type": "boolean"
        },
        "attachment": {
          "type": "string"
        },
        "createDate": {
          "type": "number"
        },
        "updateDate": {
          "type": "number"
        }
      }
    },
...
  "Plan":  {
    "required": [
      "name",
      "email",
      "password",
      "isSuperAdmin",
      "permissions"
    ],
    "properties": {
      "planId": {
        "type": "number"
      },
      "name": {
        "type": "string"
      },
      "createdBy": {
        "type": "string"
      },
      "price": {
        "type": "number"
      },
      "books":  {
        "type": "array",
        "items": {
          "type": "object"
        }
      },
      "duration": {
        "type": "number"
      },
      "description": {
        "type": "string"
      },
      "active": {
        "type": "boolean"
      }
    }
  },
  "UpdatePlan":  {
    "required": [
      "_id"
    ],
    "properties": {
      "_id": {
        "type": "string"
      },
      "planId": {
        "type": "number"
      },
      "name": {
        "type": "string"
      },
      "createdBy": {
        "type": "string"
      },
      "price": {
        "type": "number"
      },
      "books":  {
        "type": "array",
        "items": {
          "type": "object"
        }
      },
      "duration": {
        "type": "number"
      },
      "description": {
        "type": "string"
      },
      "active": {
        "type": "boolean"
      }
    }
  },
...
  }
}